Add commands to set_whitelist management command
- Refactor set_whitelist to use a command pattern.
- Add commands for basic manipulation of the whitelist.
- The "bulk" command replaces the earlier set_whitelist functionality.

view: View the whitelist
add <subject>: Add a subject to the whitelist (Also see ./manage.py cert
whitelist)
remove <subject>: Remove a subject from the whitelist
bulk <file>: Create whitelist from file (one subject per line, blank and
comment lines ignored)

+++ d1_mn_generic/src/gmn/app/management/commands/whitelist.py
@@ -17,12 +17,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""Update the CRUD whitelist
-
-When running in production, GMN requires that subjects that wish to create,
-update or delete science objects are registered in a whitelist. This command
-manages the whitelist.
+"""Manage the whitelist for access to create, update and delete operations
 """
 
 # Stdlib.
@@ -31,12 +26,26 @@ import logging
 # Django.
 import django.core.management.base
 
+# D1.
+import d1_common.types.exceptions
+import d1_common.util
+
 # App.
+import app.middleware.session_cert
 import app.models
+import util
 
 
 class Command(django.core.management.base.BaseCommand):
-  help = 'Set the whitelist for create, update and delete operations'
+  help = 'Manage the whitelist for access to create, update and delete operations'
+
+  missing_args_message = (
+    '<command> must be one of:\n'
+    'view: View the whitelist\n'
+    'add <subject>: Add a subject to the whitelist (Also see ./manage.py cert whitelist)\n'
+    'remove <subject>: Remove a subject from the whitelist\n'
+    'bulk <file>: Create whitelist from file (one subject per line, blank and # lines ignored)\n'
+  )
 
   def add_arguments(self, parser):
     parser.add_argument(
@@ -46,27 +55,86 @@ class Command(django.core.management.base.BaseCommand):
       help='debug level logging',
     )
     parser.add_argument(
-      'path', type=str, help='path to file containing subjects to whitelist'
+      'command',
+      help='valid commands: view, add, remove, bulk',
+    )
+    parser.add_argument(
+      'command_arg',
+      default=None,
+      nargs='?',
+      help='argument for command',
     )
 
   def handle(self, *args, **options):
     util.log_setup(options['debug'])
-    logging.info(
-      u'Running management command: {}'.format(util.get_command_name())
-    )
-    util.abort_if_other_instance_is_running()
-    num_subjects = self.set_whitelist(options['path'])
-    logging.info(u'Whitelisted {} subjects'.format(num_subjects))
+    if options['command'] not in ('view', 'add', 'remove', 'bulk'):
+      logging.info(self.missing_args_message)
+      return
+    try:
+      self._handle(options['command'], options['command_arg'])
+    except d1_common.types.exceptions.DataONEException as e:
+      raise django.core.management.base.CommandError(str(e))
 
-  def set_whitelist(self, whitelist_path):
-    with open(whitelist_path) as f:
+  def _handle(self, command_str, command_arg_str):
+    if command_str == 'view':
+      self._view()
+    elif command_str == 'add':
+      self._add(command_arg_str)
+    elif command_str == 'remove':
+      self._remove(command_arg_str)
+    elif command_str == 'bulk':
+      self._bulk(command_arg_str)
+    else:
+      raise django.core.management.base.CommandError('Unknown command')
+
+  def _view(self):
+    logging.info(u'Subjects in whitelist:')
+    for whitelist_model in app.models.WhitelistForCreateUpdateDelete.objects.all():
+      logging.info('  {}'.format(whitelist_model.subject.subject))
+
+  def _add(self, subject_str):
+    if subject_str is None:
+      raise django.core.management.base.CommandError(
+        u'Please specify a subject to add',
+      )
+    if util.is_subject_in_whitelist(subject_str):
+      raise django.core.management.base.CommandError(
+        u'Subject already in whitelist: {}'.format(subject_str)
+      )
+    app.models.whitelist_for_create_update_delete(subject_str)
+    logging.info(u'Added subject to whitelist: {}'.format(subject_str))
+
+  def _remove(self, subject_str):
+    if subject_str is None:
+      raise django.core.management.base.CommandError(
+        u'Please specify a subject to remove',
+      )
+    if not util.is_subject_in_whitelist(subject_str):
+      raise django.core.management.base.CommandError(
+        u'Subject is not in whitelist: {}'.format(subject_str)
+      )
+    app.models.WhitelistForCreateUpdateDelete.objects.filter(
+      subject=app.models.subject(subject_str)
+    ).delete()
+    logging.info(
+      u'Removed subject from whitelist: {}'.format(subject_str)
+    )
+
+  def _bulk(self, whitelist_file_path):
+    if whitelist_file_path is None:
+      raise django.core.management.base.CommandError(
+        u'Please specify path to whitelist file',
+      )
+    subject_cnt = 0
+    with open(whitelist_file_path) as f:
       app.models.WhitelistForCreateUpdateDelete.objects.all().delete()
-      cnt = 0
       for subject_str in f:
         subject_str = subject_str.strip()
         if subject_str == '' or subject_str.startswith('#'):
           continue
         app.models.whitelist_for_create_update_delete(subject_str)
-        cnt += 1
+        subject_cnt += 1
+    logging.info(
+      u'Created new whitelist with {} subjects'.format(subject_cnt)
+    )
 
-    return cnt
